Tidying up code, trying to organise system a bit.
