--- Scripts/Login.js
@@ -1,2 +1,0 @@
-// A script for a basic login function...
-//TODO LATER AFTER FUNDEMENTALS ARE DONE
