+++ home.html
@@ -23,9 +23,11 @@
         </ul>
     </div>
 
+    <button id="TestButt">Button</button>
+
     <!--Need to work on the CSS for this.-->
     <div class="HomeContent">
-        <div id="Blocks"></div>
+        <div id="Blocks"\>
     </div>
 </body>
