+++ style.css
@@ -132,7 +132,7 @@ body {
     }
 
     .MarketContent #Blocks div{
-        width: 200px;
+        width: 300px;
         height: 100px;
         /*Can't get this working right at the moment. Can't be bothered at the moment.*/
         margin: 2px;
@@ -146,7 +146,7 @@ body {
     }
 
     .MarketContent #Blocks p {
-        width: 100px;
+        width: 150px;
         height: 100px;
         margin: 0;
     }
